refactor: fix folder name typo
