+++ src/Modules/Resources/Tests/IntegrationTests/WorldNodes/DecreaseExtractionRateTests.cs
@@ -1,12 +1,13 @@
 ﻿using SatisfactoryPlanner.BuildingBlocks.Application;
+using SatisfactoryPlanner.Modules.Resources.Application.WorldNodes.DecreaseExtractionRate;
 using SatisfactoryPlanner.Modules.Resources.Application.WorldNodes.GetWorldNodeDetails;
 using SatisfactoryPlanner.Modules.Resources.Application.WorldNodes.IncreaseExtractionRate;
 using SatisfactoryPlanner.Modules.Resources.IntegrationTests.SeedWork;
 
-namespace SatisfactoryPlanner.Modules.Resources.IntegrationTests.WorldsNodes
+namespace SatisfactoryPlanner.Modules.Resources.IntegrationTests.WorldNodes
 {
     [TestFixture]
-    public class IncreaseExtractionRateTests : TestBase
+    public class DecreaseExtractionRateTests : TestBase
     {
         // Happy path tests
         [Test]
@@ -15,10 +16,24 @@ namespace SatisfactoryPlanner.Modules.Resources.IntegrationTests.WorldsNodes
             var (worldId, nodeId) = await new TappedWorldNodeFixture().Create(ResourcesModule);
 
             await ResourcesModule.ExecuteCommandAsync(new IncreaseExtractionRateCommand(worldId, nodeId, 21));
+            await ResourcesModule.ExecuteCommandAsync(new DecreaseExtractionRateCommand(worldId, nodeId, 10));
 
             var postTapNodeDetails =
                 await ResourcesModule.ExecuteQueryAsync(new GetWorldNodeDetailsQuery(worldId, nodeId));
-            postTapNodeDetails.ExtractionRate.Should().Be(21);
+            postTapNodeDetails.ExtractionRate.Should().Be(10);
+        }
+
+        [Test]
+        public async Task WhenExtractionRateIsZero_IsSuccessful()
+        {
+            var (worldId, nodeId) = await new TappedWorldNodeFixture().Create(ResourcesModule);
+
+            await ResourcesModule.ExecuteCommandAsync(new IncreaseExtractionRateCommand(worldId, nodeId, 21));
+            await ResourcesModule.ExecuteCommandAsync(new DecreaseExtractionRateCommand(worldId, nodeId, 0));
+
+            var postTapNodeDetails =
+                await ResourcesModule.ExecuteQueryAsync(new GetWorldNodeDetailsQuery(worldId, nodeId));
+            postTapNodeDetails.ExtractionRate.Should().Be(0);
         }
 
         // CommandValidator tests
@@ -28,7 +43,7 @@ namespace SatisfactoryPlanner.Modules.Resources.IntegrationTests.WorldsNodes
             Assert.CatchAsync<InvalidCommandException>(async () =>
             {
                 await ResourcesModule.ExecuteCommandAsync(
-                    new IncreaseExtractionRateCommand(Guid.Empty, Guid.NewGuid(), 1));
+                    new DecreaseExtractionRateCommand(Guid.Empty, Guid.NewGuid(), 1));
             });
         }
 
@@ -38,19 +53,7 @@ namespace SatisfactoryPlanner.Modules.Resources.IntegrationTests.WorldsNodes
             Assert.CatchAsync<InvalidCommandException>(async () =>
             {
                 await ResourcesModule.ExecuteCommandAsync(
-                    new IncreaseExtractionRateCommand(Guid.NewGuid(), Guid.Empty, 1));
-            });
-        }
-
-        [Test]
-        public async Task WhenExtractionRateIsZero_ThrowsInvalidCommandException()
-        {
-            var (worldId, nodeId) = await new TappedWorldNodeFixture().Create(ResourcesModule);
-
-            Assert.CatchAsync<InvalidCommandException>(async () =>
-            {
-                await ResourcesModule.ExecuteCommandAsync(
-                    new IncreaseExtractionRateCommand(worldId, nodeId, 0));
+                    new DecreaseExtractionRateCommand(Guid.NewGuid(), Guid.Empty, 1));
             });
         }
 
@@ -62,7 +65,7 @@ namespace SatisfactoryPlanner.Modules.Resources.IntegrationTests.WorldsNodes
             Assert.CatchAsync<InvalidCommandException>(async () =>
             {
                 await ResourcesModule.ExecuteCommandAsync(
-                    new IncreaseExtractionRateCommand(worldId, nodeId, -1));
+                    new DecreaseExtractionRateCommand(worldId, nodeId, -1));
             });
         }
 
@@ -76,7 +79,7 @@ namespace SatisfactoryPlanner.Modules.Resources.IntegrationTests.WorldsNodes
             Assert.CatchAsync<InvalidCommandException>(async () =>
             {
                 await ResourcesModule.ExecuteCommandAsync(
-                    new IncreaseExtractionRateCommand(differentWorldId, nodeId, -1));
+                    new DecreaseExtractionRateCommand(differentWorldId, nodeId, -1));
             });
         }
 
@@ -89,7 +92,7 @@ namespace SatisfactoryPlanner.Modules.Resources.IntegrationTests.WorldsNodes
             Assert.CatchAsync<InvalidCommandException>(async () =>
             {
                 await ResourcesModule.ExecuteCommandAsync(
-                    new IncreaseExtractionRateCommand(worldId, randomNodeId, -1));
+                    new DecreaseExtractionRateCommand(worldId, randomNodeId, -1));
             });
         }
     }
