+++ src/Modules/Resources/Tests/IntegrationTests/WorldNodes/TapWorldNodeTests.cs
@@ -4,7 +4,7 @@ using SatisfactoryPlanner.Modules.Resources.Application.WorldNodes.GetWorldNodes
 using SatisfactoryPlanner.Modules.Resources.Application.WorldNodes.TapWorldNode;
 using SatisfactoryPlanner.Modules.Resources.IntegrationTests.SeedWork;
 
-namespace SatisfactoryPlanner.Modules.Resources.IntegrationTests.WorldsNodes
+namespace SatisfactoryPlanner.Modules.Resources.IntegrationTests.WorldNodes
 {
     [TestFixture]
     public class TapWorldNodeTests : TestBase
